+++ pyproject.toml
@@ -44,7 +44,7 @@ known_third_party = [
 ]
 known_pytorch = ["torch", "torchvision"]
 known_first_party = ["pystiche"]
-known_local_folder = ["utils"]
+known_local_folder = ["utils", "tests"]
 sections= ["FUTURE", "STDLIB", "THIRDPARTY", "PYTORCH", "FIRSTPARTY", "LOCALFOLDER"]
 
 skip = [
